% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_gtf.R
\name{deal_gtf}
\alias{deal_gtf}
\title{Download a gtf file from ncbi.}
\usage{
deal_gtf(gtf_file)
}
\arguments{
\item{gtf_file}{The directory of a input gtf file.}
}
\value{
gene & protein informations.
}
\description{
Download a gtf file from ncbi and get gene & protein infomations.
}
\examples{
NA
}
