% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_gtf.R
\name{down_gtf}
\alias{down_gtf}
\title{Download a gtf file from ncbi.}
\usage{
down_gtf(No.assembly)
}
\arguments{
\item{No.assembly}{NCBI assembly number.}
}
\value{
gene & protein informations.
}
\description{
Download a gtf file from ncbi and get gene & protein infomations.
}
\examples{
down_gtf("GCF_000001405.40")
}
