% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{Get interactions nodes from STRINGdb.}
\usage{
get_nodes(
  identifiers,
  species = 9606,
  required_score = 400,
  network_type = "functional",
  show_query_node_labels = 0,
  add_nodes = 10,
  caller_identity = "Anonymous"
)
}
\arguments{
\item{identifiers}{A character vector of required parameter for multiple items, e.g. "TP53", c("TP53", "CKD2").}

\item{species}{NCBI taxon identifiers (e.g. Human is 9606, see: STRING organisms).}

\item{required_score}{threshold of significance to include an interaction, a number between 0 and 1000 (default depends on the network).}

\item{network_type}{network type: functional (default), physical.}

\item{show_query_node_labels}{when available use submitted names in the preferredName column when (0 or 1) (default:0).}

\item{add_nodes}{adds a number of proteins with to the network based on their confidence score.}

\item{caller_identity}{your identifier for us.}
}
\value{
STRINGdb interaction tsv files.
}
\description{
Get interactions nodes from STRINGdb.
}
\examples{
get_nodes(c("TP53","CDK2"),species=9606)
}
