% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cellchatdb.R
\name{make_cellchatdb}
\alias{make_cellchatdb}
\title{make a custom cell chat database.}
\usage{
make_cellchatdb(interaction_nodes, enrich = NULL, annotation = NULL)
}
\arguments{
\item{interaction_nodes}{nodes file generated from get_nodes.}

\item{enrich}{A given pathway type, such as "KEGG：map04216", or "GO", as you wish.}

\item{annotation}{A given interaction type, one of "Secreted Signaling", "ECM-Recptor", "Cell-Cell Contact", "Non-protein Signaling".}
}
\value{
a cell chat database.
}
\description{
make a custom cell chat database with a input node file generated from get_nodes.
}
\examples{
NA
}
