% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kegg.R
\name{make_kegg}
\alias{make_kegg}
\title{Get your kegg information.}
\usage{
make_kegg(organs, gtf = NULL)
}
\arguments{
\item{organs}{Abbreviation of your species (hsa for human, mmu for mice, can be find in kegg website).}

\item{gtf}{A input gtf gene info generated by deal_gtf.}
}
\value{
kegg info.
}
\description{
Get your kegg information.
}
\examples{
NA
}
