% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_gtf.R
\name{deal_gtf}
\alias{deal_gtf}
\title{Load a gtf file.}
\usage{
deal_gtf(gtf_file, protein = T)
}
\arguments{
\item{gtf_file}{The directory of a input gtf file.}

\item{protein}{Get protein or gene info.}
}
\value{
gene & protein informations.
}
\description{
Load a gtf file from ncbi and get gene & protein infomations.
}
\examples{
NA
}
