% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cellchatdb.R
\name{make_cellchatdb}
\alias{make_cellchatdb}
\title{make a custom cell chat database.}
\usage{
make_cellchatdb(interaction_nodes, enrich = NULL)
}
\arguments{
\item{interaction_nodes}{node file generated from get_nodes.}

\item{enrich}{A give pathway type, such as "KEGG", or "GO", as you wish.}
}
\value{
a cell chat database.
}
\description{
make a custom cell chat database with a input node file generated from get_nodes.
}
\examples{
NA
}
