% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_customdb.R
\name{make_customdb}
\alias{make_customdb}
\title{make a custom cell chat database.}
\usage{
make_customdb(interaction_nodes, gene_info, enrich = NULL)
}
\arguments{
\item{interaction_nodes}{node file downloaded from STRINGdb.}

\item{gene_info}{gene & protein infomations generated from down_gtf().}

\item{enrich}{A give pathway type, such as "KEGG：map04216".}
}
\value{
a cell chat database.
}
\description{
make a custom cell chat database with a input node file downloaded from STRINGdb.
}
\examples{
NA
}
